Last commit
This is the final commit for the project.
I had a lot of fun making it and I learned a lot.
I've completed everything I set out to do, and I'm ready to move on to new challenges.

This commit includes code improvements, such as cleanup, bug/syntax fixes, and an informative README.
Additionally, new settings content has been added, providing information about special keys, keybinds, and
memory expressions, along with a memory expression builder prompt.

Started working on this project because I found the idea cool and was mainly focused on making it work.
I might in the future turn it into a full-fledged website.